requirements.txt
streamlit==1.32.2
streamlit-extras==0.3.0
python-dotenv==1.0.1
PyPDF2==3.0.1
nltk==3.9.1
transformers==4.41.1
sentence-transformers==2.7.0
langchain==0.1.13
chromadb==1.0.8
typing-extensions==4.12.0
SentencePiece
python-docx
accelerate
bitsandbytes
